ASV-1912 - Extracted ListAppsPresenter
